AT-ticket204-refactorCRUD
