+++ backend/app/api/endpoints/jobs.py
@@ -3,7 +3,7 @@ from typing import List
 from fastapi import APIRouter, Depends
 from sqlalchemy.orm import Session
 
-from backend.app.crud import crud
+from backend.app.crud import jobs
 from backend.app.models import Job as JobModel
 from backend.app.schemas import Job as JobSchema
 from backend.app.schemas import JobCreate
@@ -16,7 +16,7 @@ router = APIRouter()
 @router.get("/", response_model=List[JobSchema])
 async def get_jobs(db: Session = Depends(get_db), skip: int = 0, limit: int = 100):
     """Returns all jobs from the database."""
-    return crud.get_all_jobs(db, skip=skip, limit=limit)
+    return jobs.get_all_jobs(db, skip=skip, limit=limit)
 
 
 @router.post("/", response_model=JobSchema)
@@ -29,7 +29,7 @@ async def create_job(job: JobCreate, db: Session = Depends(get_db)):
             if k in JobModel.__table__.columns.keys() and v is not None
         }
     )
-    return crud.create_job(db, db_job)
+    return jobs.create_job(db, db_job)
 
 
 @router.post("/search", response_model=List[JobSchema])
@@ -37,4 +37,4 @@ async def search_jobs(
     db: Session = Depends(get_db), q: str = None, skip: int = 0, limit: int = 100
 ):
     """Searches the database for jobs."""
-    return crud.search_jobs(db, q=q, skip=skip, limit=limit)
+    return jobs.search_jobs(db, q=q, skip=skip, limit=limit)
